Move server ParamTypesBuilder to common.type_mapping.builder
