+++ verification-common/src/type_mapping.rs
@@ -39,16 +39,22 @@ pub enum TestType {
 #[derive(new)]
 /// Defines that a service with a given name implements tests of the given [TestType], and
 /// the Conjure type can be extracted from the endpoint definition using the given [TypeForEndpointFn].
+///
+/// [TestType]: enum.TestType.html
+/// [TypeForEndpointFn]: type.TypeForEndpointFn.html
 pub struct ServiceTypeMapping<'a> {
     pub service_name: &'a str,
     pub test_type: TestType,
     pub type_for_endpoint_fn: TypeForEndpointFn,
 }
 
+/// Type alias describing the mapping from test type -> endpoint name -> a conjure type.
+pub type ParamTypes = HashMap<TestType, HashMap<EndpointName, ResolvedType>>;
+
 pub fn resolve_types<'a, 'b>(
     ir: &'a Conjure,
     type_by_service: &'a [ServiceTypeMapping<'b>],
-) -> Box<HashMap<TestType, HashMap<EndpointName, ResolvedType>>> {
+) -> Box<ParamTypes> {
     // Resolve endpoint -> type mappings eagerly
     let mut param_types = Box::new(HashMap::new());
     type_by_service.iter().for_each(
@@ -100,3 +106,32 @@ pub fn return_type(endpoint_def: &ir::EndpointDefinition) -> &ir::Type {
 }
 
 pub type TypeForEndpointFn = fn(&ir::EndpointDefinition) -> &ir::Type;
+
+/// Builder for easy construction of of the return type of [resolve_types](fn.resolve_types.html).
+pub mod builder {
+    use super::*;
+
+    type ParamTypes = HashMap<TestType, HashMap<EndpointName, ResolvedType>>;
+
+    #[derive(Default)]
+    pub struct ParamTypesBuilder(ParamTypes);
+
+    impl ParamTypesBuilder {
+        pub fn add(
+            &mut self,
+            tt: TestType,
+            endpoint_name: EndpointName,
+            resolved_type: ResolvedType,
+        ) -> &mut Self {
+            self.0
+                .entry(tt)
+                .or_default()
+                .insert(endpoint_name, resolved_type);
+            self
+        }
+
+        pub fn build(self) -> ParamTypes {
+            self.0
+        }
+    }
+}
